refactor: migrar servicios a UoW, mejorar schemas y corregir hooks
- Migrar admin/service y catalogo/service al patrón UoW
- Usar schemas Pydantic (RolRead, UnidadMedidaRead, etc.) en vez de dicts manuales
- Simplificar pedidos/router delegando lógica al service
- Corregir hooks de WebSocket en ambos frontends
- Agregar BusinessRuleError en pagos y uploads routers
- Mejorar tipado en schemas de ingredientes y productos

+++ frontend-admin/src/schemas/ingrediente.schema.ts
@@ -4,7 +4,7 @@ export const ingredienteFormSchema = z.object({
   nombre: z.string().min(1, 'El nombre es requerido').min(3, 'Minimo 3 caracteres'),
   descripcion: z.string().default(''),
   es_alergeno: z.boolean().default(false),
-  stock_cantidad: z.number().int().min(0, 'El stock no puede ser negativo').default(0),
+  stock_cantidad: z.number().min(0, 'El stock no puede ser negativo').default(0),
   unidad_medida_id: z.number().int().min(1, 'La unidad de medida es requerida'),
 });
